fix: updates for iOS 11.4 & Xcode 9.4
The one method of SCNGeometrySource that we need most is hidden behind a
silly macro, so I define an extension which allows Swift to find it.

This commit also fixes up some of the swift compile errors. I'm not a
Swift hacker, so I've just followed what the compiler suggested that I
do. ? and ! after variables?!?!?!?!??!

Runs successfully on my iPhone 7 plus running iOS 11.4.